meson: Add a simple cross file for mingw32 on Ubuntu.
Unfortunately this doesn't work, because when Meson finds the first dependency

  dep_glew = dependency('glew', version : '>=1.5.4')

it immediately fails with

  Pkg-config binary missing from cross file

It doesn't make sense to require pkgconfig when building for Windows, as
nobody uses it.

CMake doesn't require pkg-config.  Instead it allows one to pupolate
dependencies automatically via "Cache" files.  For example, when I build
Mesa demos with CMake for Windows I have cache file containing:

   set (GLEW_INCLUDE_DIR "/path/to/glew/include" CACHE PATH "" FORCE)
   set (GLEW_glew_LIBRARY "/path/to/glew/lib/libglew32.a" CACHE FILEPATH "" FORCE)

and so on for other dependencies.  But I don't see how to accomplish
this with Meson...