Example using cross properties
This is an example of using cross properties to set where certain
libraries are.

This also sets the pkgconfig variable in the ubuntu cross file (which
also works for Archlinux). This is required to make rebuilds work
correctly, but is likely a bug in meson. Setting it to 'false' seems to
make things work.